% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeeDeeExperiment.R
\name{add_dea}
\alias{add_dea}
\title{dde : the deedee exp object
dea: the named list, as for the main constructor
Title}
\usage{
add_dea(dde, dea)
}
\arguments{
\item{dde}{todo}

\item{dea}{todo}
}
\value{
todo
}
\description{
dde : the deedee exp object
dea: the named list, as for the main constructor
Title
}
\examples{
# todo
}
