% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee-data.R
\docType{data}
\name{dde_macrophage}
\alias{dde_macrophage}
\title{A sample \code{DeeDeeExperiment} object}
\description{
A sample \code{DeeDeeExperiment} object, built around a \code{DESeqDataSet} from the
\code{macrophage} package, containing four DESeq2 contrasts as differential
expression results.
}
\details{
The underlying \code{DESeqDataSet} compares macrophages under four
conditions: naive, IFNg-stimulated, Salmonella-infected, and both
IFNg-stimulated and Salmonella-infected. The \code{dea} slot of this object
contains four contrasts:
\itemize{
\item \code{IFNg_naive}: IFNg vs. naive
\item \code{IFNg_both}: IFNg + Salmonella vs. IFNg
\item \code{Salm_naive}: Salmonella vs. naive
\item \code{Salm_both}: IFNg + Salmonella vs. Salmonella
}

The code to create this object can be found in \verb{/inst/scripts/generate_data.R}
in the DeeDee package.
}
\references{
Alasoo, et al. "Shared genetic effects on chromatin and gene
expression indicate a role for enhancer priming in immune response",
Nature Genetics, January 2018 doi: 10.1038/s41588-018-0046-7.
}
