% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeeDee-app.R
\name{ddedde_app}
\alias{ddedde_app}
\title{DeeDee App}
\usage{
ddedde_app(deedee_obj = NULL, dde = NULL)
}
\arguments{
\item{deedee_obj}{An object of the class DeeDeeExperiment to be analyzed.}

\item{dde}{An object of the class DeeDeeExperiment to be analyzed.}
}
\value{
A shiny app
}
\description{
\code{deedee_app} opens the DeeDee Shiny web application, combining
the functionalities of all other DeeDee functions with a user-friendly
graphical user interface.
}
\examples{
data("de_named_list", package = "DeeDee")
library("SummarizedExperiment")

rd_macrophage <- DataFrame(
  gene_id = rownames(del$ifng_vs_naive))
rownames(rd_macrophage) <- rownames(del$ifng_vs_naive)
se_macrophage_noassays <- SummarizedExperiment(
  assays = SimpleList(),
  rowData = rd_macrophage
)
dde <- DeeDeeExperiment(
  se_macrophage_noassays,
  de_results = del
)
dde

if (interactive()) {
  ddedde_app(dde = dde)
}

# one can provide the DeeDeeExperiment object also at runtime
if (interactive()) {
  ddedde_app()
}
}
