% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeeDee_bars.R
\name{ddedde_bars}
\alias{ddedde_bars}
\title{Title}
\usage{
ddedde_bars(dde)
}
\arguments{
\item{dde}{todo}
}
\value{
todo
}
\description{
Title
}
\examples{
data("de_named_list", package = "DeeDee")
library(SummarizedExperiment)

rd_macrophage <- DataFrame(
  gene_id = rownames(del$ifng_vs_naive))
rownames(rd_macrophage) <- rownames(del$ifng_vs_naive)
se_macrophage_noassays <- SummarizedExperiment(
  assays = SimpleList(),
  rowData = rd_macrophage
)
dde <- DeeDeeExperiment(
  se_macrophage_noassays,
  del
)
dde
}
