% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeeDee_funcs.R
\name{ddedde_cat}
\alias{ddedde_cat}
\title{deedee_heatmap(dd_list_original,
pthresh = 0.05, show_first = 25,
show_gene_names = FALSE, dist = "euclidean",
clust = "average", show_na = FALSE
)
deedee_cat(dd_list_original, ref = 1, maxrank = 1000, mode = "up", pthresh = 0.05)
Title}
\usage{
ddedde_cat(dde, ref = 1, maxrank = 1000, mode = "up", pthresh = 0.05)
}
\arguments{
\item{dde}{todo}

\item{ref}{todo}

\item{maxrank}{todo}

\item{mode}{todo}

\item{pthresh}{todo}
}
\value{
todo
}
\description{
deedee_heatmap(dd_list_original,
pthresh = 0.05, show_first = 25,
show_gene_names = FALSE, dist = "euclidean",
clust = "average", show_na = FALSE
)
deedee_cat(dd_list_original, ref = 1, maxrank = 1000, mode = "up", pthresh = 0.05)
Title
}
\examples{
data("de_named_list", package = "DeeDee")
library("SummarizedExperiment")

rd_macrophage <- DataFrame(
  gene_id = rownames(del$ifng_vs_naive))
rownames(rd_macrophage) <- rownames(del$ifng_vs_naive)
se_macrophage_noassays <- SummarizedExperiment(
  assays = SimpleList(),
  rowData = rd_macrophage
)
dde <- DeeDeeExperiment(
  se_macrophage_noassays,
  del
)
dde
}
