% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeeDee_qq.R
\name{ddedde_qq}
\alias{ddedde_qq}
\title{Title}
\usage{
ddedde_qq(
  dde,
  select1 = 1,
  select2 = 2,
  color_by = "pval1",
  as_line = FALSE,
  pthresh = 0.05
)
}
\arguments{
\item{dde}{todo}

\item{select1}{todo}

\item{select2}{todo}

\item{color_by}{todo}

\item{as_line}{todo}

\item{pthresh}{todo}
}
\value{
todo
}
\description{
Title
}
\examples{
data("de_named_list", package = "DeeDee")
library("SummarizedExperiment")

rd_macrophage <- DataFrame(
  gene_id = rownames(del$ifng_vs_naive))
rownames(rd_macrophage) <- rownames(del$ifng_vs_naive)
se_macrophage_noassays <- SummarizedExperiment(
  assays = SimpleList(),
  rowData = rd_macrophage
)
dde <- DeeDeeExperiment(
  se_macrophage_noassays,
  de_results = del
)
dde

ddedde_qq(dde)

}
