% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeeDee_scatter.R
\name{ddedde_scatter}
\alias{ddedde_scatter}
\title{Title}
\usage{
ddedde_scatter(
  dde,
  select1 = 1,
  select2 = 2,
  color_by = "pval1",
  pthresh = 0.05
)
}
\arguments{
\item{dde}{todo}

\item{select1}{todo}

\item{select2}{todo}

\item{color_by}{todo}

\item{pthresh}{todo}
}
\value{
todo
}
\description{
Title
}
\examples{
data("de_named_list", package = "DeeDee")
library("SummarizedExperiment")

rd_macrophage <- DataFrame(
  gene_id = rownames(del$ifng_vs_naive))
rownames(rd_macrophage) <- rownames(del$ifng_vs_naive)
se_macrophage_noassays <- SummarizedExperiment(
  assays = SimpleList(),
  rowData = rd_macrophage
)
dde <- DeeDeeExperiment(
  se_macrophage_noassays,
  del
)
dde
ddedde_scatter(dde, 1, 3, color_by = "pval2")
ddedde_scatter(dde, 1, 2)
ddedde_scatter(dde, 1, 4)
ddedde_scatter(dde, 3, 2)
}
