% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeeDee_funcs.R
\name{ddedde_summary}
\alias{ddedde_summary}
\title{ddedde_qqmult(dde,
pthresh = 0.05,
ref = 1)}
\usage{
ddedde_summary(
  dde,
  template = system.file("extdata", "summary_template.Rmd", package = "DeeDee"),
  output_path = "DeeDee_Summary.html",
  overwrite = FALSE,
  pthresh = 0.05,
  scatter_select1 = 1,
  scatter_select2 = 2,
  scatter_color_by = "pval1",
  heatmap_show_first = 25,
  heatmap_show_gene_names = FALSE,
  heatmap_dist = "euclidean",
  heatmap_clust = "average",
  heatmap_show_na = FALSE,
  venn_mode = "both",
  upset_mode = "both_colored",
  upset_min_setsize = 10,
  qqmult_ref = 1,
  cat_ref = 1,
  cat_maxrank = 1000,
  cat_mode = "up",
  silent = FALSE,
  open_file = TRUE
)
}
\arguments{
\item{dde}{todo}

\item{template}{todo}

\item{output_path}{todo}

\item{overwrite}{todo}

\item{pthresh}{todo}

\item{scatter_select1}{todo}

\item{scatter_select2}{todo}

\item{scatter_color_by}{todo}

\item{heatmap_show_first}{todo}

\item{heatmap_show_gene_names}{todo}

\item{heatmap_dist}{todo}

\item{heatmap_clust}{todo}

\item{heatmap_show_na}{todo}

\item{venn_mode}{todo}

\item{upset_mode}{todo}

\item{upset_min_setsize}{todo}

\item{qqmult_ref}{todo}

\item{cat_ref}{todo}

\item{cat_maxrank}{todo}

\item{cat_mode}{todo}

\item{silent}{todo}

\item{open_file}{todo}
}
\value{
todo
}
\description{
deedee_qqmult(dd_list_original, pthresh = 0.05, ref = 2)
ddedde_qqmult(dde, pthresh = 0.05, ref = 2)
}
\details{
deedee_qqmult(dd_list_original, pthresh = 0.05, ref = 3)
ddedde_qqmult(dde, pthresh = 0.05, ref = 3)

deedee_qqmult(dd_list_original, pthresh = 0.05, ref = 4)
ddedde_qqmult(dde, pthresh = 0.05, ref = 4)
deedee_summary(dd_list_original)
Title
}
\examples{
data("de_named_list", package = "DeeDee")
library("SummarizedExperiment")

rd_macrophage <- DataFrame(
  gene_id = rownames(del$ifng_vs_naive))
rownames(rd_macrophage) <- rownames(del$ifng_vs_naive)
se_macrophage_noassays <- SummarizedExperiment(
  assays = SimpleList(),
  rowData = rd_macrophage
)
dde <- DeeDeeExperiment(
  se_macrophage_noassays,
  del
)
dde
}
