% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeeDee_venn.R
\name{ddedde_venn}
\alias{ddedde_venn}
\title{deedee_venn(dd_list_original, pthresh = 0.05, mode = "both")
Title}
\usage{
ddedde_venn(dde, mode = "both", pthresh = 0.05)
}
\arguments{
\item{dde}{todo}

\item{mode}{todo}

\item{pthresh}{todo}
}
\value{
todo
}
\description{
deedee_venn(dd_list_original, pthresh = 0.05, mode = "both")
Title
}
\examples{
data("de_named_list", package = "DeeDee")
library("SummarizedExperiment")

rd_macrophage <- DataFrame(
  gene_id = rownames(del$ifng_vs_naive))
rownames(rd_macrophage) <- rownames(del$ifng_vs_naive)
se_macrophage_noassays <- SummarizedExperiment(
  assays = SimpleList(),
  rowData = rd_macrophage
)
dde <- DeeDeeExperiment(
  se_macrophage_noassays,
  de_results = del
)
dde
}
