% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeeDee-data.R
\docType{data}
\name{de_named_list}
\alias{de_named_list}
\title{de_named_list}
\value{
A named list of DE results, in their original format
}
\description{
A named list of DE results, in their original format (from DESeq2, edgeR or limma)
}
\details{
documented creation in the \code{create_dataset_example.R} script in the
\code{scripts} package folder
}
