% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeeDee_bars.R
\name{deedee_bars}
\alias{deedee_bars}
\title{Title}
\usage{
deedee_bars(dde, p_thresh = 0.05, show_DEnumbers = TRUE)
}
\arguments{
\item{dde}{A \link{DeeDeeExperiment} object.}

\item{p_thresh}{Numeric value, corresponding to the threshold used to call a
feature as differentially expressed.}

\item{show_DEnumbers}{Logical value;}
}
\value{
A \code{ggplot} plot object, summarizing the amount of DE genes in each
comparison performed.
}
\description{
Title
}
\examples{
data("de_named_list", package = "DeeDee")
library(SummarizedExperiment)

rd_macrophage <- DataFrame(
  gene_id = rownames(del$ifng_vs_naive))
rownames(rd_macrophage) <- rownames(del$ifng_vs_naive)
se_macrophage_noassays <- SummarizedExperiment(
  assays = SimpleList(),
  rowData = rd_macrophage
)
dde <- DeeDeeExperiment(
  se_macrophage_noassays,
  de_results = del
)
dde

deedee_bars(dde)
}
