% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_cat.R
\name{deedee_cat}
\alias{deedee_cat}
\title{DeeDee Concordance At the Top Plot}
\usage{
deedee_cat(data, ref = 1, maxrank = 1000, mode = "up", pthresh = 0.05)
}
\arguments{
\item{data}{instance of the DeeDeeExperiment class}

\item{ref}{index of the contrast in data to be used as reference contrast
(default = 1)}

\item{maxrank}{highest rank that should be displayed (default = 1000)}

\item{mode}{sort by highest logFC (\code{up}, default), lowest logFC (\code{down}) or
greatest deviation from zero (\code{both})}

\item{pthresh}{threshold for p-values to be in-/excluded (default = 0.05)}
}
\value{
ggplot object (plottable with show()/print())
}
\description{
\code{deedee_cat} creates a Concordance at the Top plot depicting
the concordance of genes in the first n elements of the logFC-ranked lists
for the given (non-reference) input contrasts against a chosen reference
contrast from the input list.
}
\examples{

data(dde_macrophage, package = "DeeDee")
deedee_cat(dde_macrophage, maxrank = 500)

}
