% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_cat2.R
\name{deedee_cat2}
\alias{deedee_cat2}
\title{Title}
\usage{
deedee_cat2(data, pthresh = 0.05, maxrank = 1000)
}
\arguments{
\item{data}{(named) list of results from deedee_prepare}

\item{pthresh}{threshold for p-values to be in-/excluded (default = 0.05)}

\item{maxrank}{highest rank that should be displayed (default = 1000)}
}
\value{
ggplot object (plottable with show()/print())
}
\description{
"deedee_CAT2" creates a Concordance at the Top plot depicting
the concordance of genes in the first n elements of the logFC-ranked lists
for the given input datasets.
}
\examples{
deedee_CAT2(list(one = inp1, two = inp2, three = inp3, four = inp4),
                     pthresh = 0.08,
                     select1 = 1,
                     select2 = 4)

}
