% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeeDee_heatmap.R
\name{deedee_deheat}
\alias{deedee_deheat}
\title{deedee_heatmap(dd_list_original,
pthresh = 0.05, show_first = 25,
show_gene_names = FALSE, dist = "euclidean",
clust = "average", show_na = FALSE
)
Title}
\usage{
deedee_deheat(
  dde,
  assay_name = "vst",
  pvalue_threshold = 0.05,
  logfc_threshold = 5,
  custom_genelist = NULL
)
}
\arguments{
\item{dde}{TODO}

\item{assay_name}{TODO}

\item{pvalue_threshold}{TODO}

\item{logfc_threshold}{TODO}

\item{custom_genelist}{TODO}
}
\value{
TODO
}
\description{
deedee_heatmap(dd_list_original,
pthresh = 0.05, show_first = 25,
show_gene_names = FALSE, dist = "euclidean",
clust = "average", show_na = FALSE
)
Title
}
\examples{
# TODO
NULL ## TODO: this needs to be fully in
}
