% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_fea.R
\name{deedee_fea}
\alias{deedee_fea}
\title{DeeDee Functional Enrichment (on Scatter)}
\usage{
deedee_fea(
  data,
  select1 = 1,
  select2 = 2,
  genes,
  species = "Homo_sapiens",
  ontology = "BP"
)
}
\arguments{
\item{data}{instance of the DeeDeeExperiment class}

\item{select1}{index of first data-list element to be used (default = 1)}

\item{select2}{index of second data-list element to be used (default = 2)}

\item{genes}{character vector of gene symbols to test for enrichment
(e.g. the genes selected from a \code{deedee_scatter()} plot)}

\item{species}{species the data comes from (possible values = \code{Homo_sapiens}
(default), \code{Mus_musculus})}

\item{ontology}{which GO domain to test (possible values = \code{BP} (default),
\code{MF}, \code{CC})}
}
\value{
a data.frame of GO enrichment results
}
\description{
\code{deedee_fea} runs \code{topGO} analysis (via \code{mosdef::run_topGO()})
on a set of genes, e.g. genes selected interactively from the
\code{deedee_scatter()} plot. The universe for the test is every gene tested in
\emph{both} selected contrasts.
}
