% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_overlap.R
\name{deedee_overlap}
\alias{deedee_overlap}
\title{DeeDee Overlap Gene List}
\usage{
deedee_overlap(data, mode = "both", pthresh = 0.05, select = NULL)
}
\arguments{
\item{data}{instance of the DeeDeeExperiment class}

\item{mode}{show all overlapping DE genes (\code{both}, default),
only conjointly up-regulated (\code{up})
or only conjointly down-regulated (\code{down}) genes}

\item{pthresh}{threshold for p-values to be in-/excluded (default = 0.05)}

\item{select}{vector of DEA slot indexes to use for plotting, must be at least 2}
}
\value{
data.frame with one row per overlapping gene
}
\description{
\code{deedee_overlap} computes the differentially expressed genes
shared across all selected contrasts, filtered by p-value threshold and
regulation direction.
}
\examples{

data(dde_macrophage, package = "DeeDee")
deedee_overlap(dde_macrophage, mode = "both", pthresh = 0.05)

}
