% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_summary.R
\name{deedee_summary}
\alias{deedee_summary}
\title{DeeDee Summary}
\usage{
deedee_summary(
  data,
  output_path = "DeeDee_Summary.html",
  overwrite = FALSE,
  pthresh = 0.05,
  scatter_select1 = 1,
  scatter_select2 = 2,
  scatter_color_by = "pval1",
  heatmap_show_first = 25,
  heatmap_show_gene_names = FALSE,
  heatmap_dist = "euclidean",
  heatmap_clust = "average",
  heatmap_show_na = FALSE,
  venn_mode = "both",
  upset_mode = "both_colored",
  upset_min_setsize = 10,
  qqmult_ref = 1,
  cat_ref = 1,
  cat_maxrank = 1000,
  cat_mode = "up",
  silent = FALSE,
  open_file = TRUE
)
}
\arguments{
\item{data}{instance of the DeeDeeExperiment class;
supported legacy: named list of results from deedee_prepare()}

\item{output_path}{the path to save the resulting report in. must end with
a filename.html (default = "DeeDee_Summary.html" in the
working directory)}

\item{overwrite}{logical value specifying if the output is supposed to
overwrite a potential existing file at the location of
output_path (default = FALSE)}

\item{pthresh}{threshold for p-values to be in-/excluded (default = 0.05)}

\item{scatter_select1}{index of first data-list element to be used
(default = 1)}

\item{scatter_select2}{index of second data-list element to be used
(default = 2)}

\item{scatter_color_by}{indicates which set of values the output should be
colored by (possible values = \code{pval1} (default),
\code{pval2})}

\item{heatmap_show_first}{indicating the number of genes depicted
(default = 25)}

\item{heatmap_show_gene_names}{boolean, show row names next to heatmap
(default = FALSE)}

\item{heatmap_dist}{select the distance measure (\code{euclidean}, \code{manhattan},
\code{pearson}, \code{spearman})}

\item{heatmap_clust}{select the clustering method (\code{single}, \code{complete},
\code{average}, \code{centroid})}

\item{heatmap_show_na}{boolean, include genes with NA values in heatmap
(default = FALSE)}

\item{venn_mode}{show all overlapping DE genes (\code{both}, default),
only conjointly up-regulated (\code{up})
or only conjointly down-regulated (\code{down}) genes}

\item{upset_mode}{show all overlapping DE genes (\code{both}),
all overlapping genes colored by DE direction
(\code{both_colored}, default),
only conjointly up-regulated (\code{up})
or only conjointly down-regulated (\code{down}) genes}

\item{upset_min_setsize}{the minimum size of intersections to be displayed in
the UpSet plot (default = 10)}

\item{qqmult_ref}{index of the contrast in data to be used as reference
contrast (default = 1)}

\item{cat_ref}{index of the contrast in data to be used as reference contrast
(default = 1)}

\item{cat_maxrank}{highest rank that should be displayed (default = 1000)}

\item{cat_mode}{sort by highest logFC (\code{up}, default), lowest logFC (\code{down})
or greatest deviation from zero (\code{both})}

\item{silent}{logical, specifying if success message shall be suppressed
(default = FALSE)}

\item{open_file}{logical, specifying if the output file is supposed to be
opened after successful creation (default = TRUE)}
}
\value{
Creates a html document containing the results of running the DeeDee
functions on your input data and params. TODO where
}
\description{
DeeDee Summary
}
\examples{

data(dde_macrophage, package = "DeeDee")
deedee_summary(dde_macrophage)

}
