% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_upset.R
\name{deedee_upset}
\alias{deedee_upset}
\title{DeeDee UpSet Plot}
\usage{
deedee_upset(data, mode = "both_colored", min_setsize = 10, pthresh = 0.05)
}
\arguments{
\item{data}{instance of the DeeDeeExperiment class}

\item{mode}{show all overlapping DE genes (\code{both}),
all overlapping genes colored by DE direction (\code{both_colored},
default),
only conjointly up-regulated (\code{up})
or only conjointly down-regulated (\code{down}) genes}

\item{min_setsize}{the minimum size of intersections to be displayed in the
UpSet plot (default = 10)}

\item{pthresh}{threshold for p-values to be in-/excluded (default = 0.05)}
}
\value{
upset element (plottable with show()/print())
}
\description{
\code{deedee_upset} creates an UpSet plot depicting the overlaps of
differentially expressed genes in the input datasets.
}
\examples{

data(dde_macrophage, package = "DeeDee")
deedee_upset(dde_macrophage, mode = "both_colored")

}
