% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_venn.R
\name{deedee_venn}
\alias{deedee_venn}
\title{DeeDee Venn Diagram}
\usage{
deedee_venn(data, mode = "both", pthresh = 0.05, select = NULL)
}
\arguments{
\item{data}{instance of the DeeDeeExperiment class}

\item{mode}{show all overlapping DE genes (\code{both}, default),
only conjointly up-regulated (\code{up})
or only conjointly down-regulated (\code{down}) genes}

\item{pthresh}{threshold for p-values to be in-/excluded (default = 0.05)}

\item{select}{vector of DEA slot indexes to use for plotting, must be 2 to 4}
}
\value{
ggplot object (plottable with show()/print())
}
\description{
\code{deedee_venn} creates a Venn diagram depicting the overlaps of
differentially expressed genes in the input datasets.
}
\examples{

data(dde_macrophage, package = "DeeDee")
deedee_venn(dde_macrophage, mode = "both", pthresh = 0.05)

}
