% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeeDeeExperiment-methods.R
\name{DeeDeeExperiment-dea}
\alias{DeeDeeExperiment-dea}
\alias{dea}
\alias{dea,DeeDeeExperiment-method}
\title{Methods for deedee exps}
\usage{
\S4method{dea}{DeeDeeExperiment}(x)
}
\arguments{
\item{x}{deex}
}
\value{
Return value varies depending on method, as described below.
}
\description{
Todo
}
