% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeeDeeExperiment-methods.R
\name{DeeDeeExperiment-methods}
\alias{DeeDeeExperiment-methods}
\alias{dea}
\alias{dea<-}
\alias{add_dea}
\alias{remove_dea}
\alias{get_dea_df}
\alias{get_dea_list}
\alias{dea,DeeDeeExperiment-method}
\alias{dea<-,DeeDeeExperiment-method}
\alias{add_dea,DeeDeeExperiment,list-method}
\alias{remove_dea,DeeDeeExperiment,character-method}
\alias{get_dea_df,DeeDeeExperiment,character-method}
\alias{get_dea_list,DeeDeeExperiment-method}
\title{Methods for \link{DeeDeeExperiment} objects}
\usage{
\S4method{dea}{DeeDeeExperiment}(x)

\S4method{dea}{DeeDeeExperiment}(x) <- value

\S4method{add_dea}{DeeDeeExperiment,list}(x, dea)

\S4method{remove_dea}{DeeDeeExperiment,character}(x, dea_name)

\S4method{get_dea_df}{DeeDeeExperiment,character}(x, dea_name)

\S4method{get_dea_list}{DeeDeeExperiment}(x)
}
\arguments{
\item{x}{A \code{\link[=DeeDeeExperiment]{DeeDeeExperiment()}} object}

\item{value}{Replacement value for replacement methods.}

\item{dea}{A named list of DE results, in any of the formats supported by
the \code{DeeDee} package (currently: results from DESeq2, edgeR, limma).}

\item{dea_name}{Character value, specifying the name of the DE analysis to
get or remove}
}
\value{
Return value varies depending on the individual methods, as described
below.
}
\description{
The \code{\link[=DeeDeeExperiment]{DeeDeeExperiment()}} class provides a family of methods to get
and set DE-related information in \link{DeeDeeExperiment} objects.
}
\details{
\itemize{
\item \code{dea} and \verb{dea<-} are the methods to get and set the \code{dea} information as a
whole. These methods return \code{DeeDeeExperiment} objects.
\item \code{add_dea} and \code{remove_dea} are used to respectively add or remove DE-results
items. These methods also return \code{DeeDeeExperiment} objects, with updated
content in the \code{dea} slot.
\item \code{get_dea_df} and \code{get_dea_list} retrieve the \code{dea} information and provide
this as a \code{DataFrame} object (for a specific analysis) or as a list, with one
element for each reported analysis.
\item \code{show} is the method to nicely print out the information of a \code{DeeDeeExperiment}
object.
}
}
\examples{
data("de_named_list", package = "DeeDee")
library("SummarizedExperiment")

rd_macrophage <- DataFrame(
  gene_id = rownames(de_named_list$ifng_vs_naive))
rownames(rd_macrophage) <- rownames(de_named_list$ifng_vs_naive)
se_macrophage_noassays <- SummarizedExperiment(
  assays = SimpleList(),
  rowData = rd_macrophage
)

# creating a `DeeDeeExperiment`
dde <- DeeDeeExperiment(
  se_macrophage_noassays,
  de_results = de_named_list
)
dde

new_del <- list(
  ifng2 = de_named_list$ifng_vs_naive,
  ifngsalmo2 = de_named_list$ifngsalmo_vs_naive
)

# add a new (set of) DE result(s)
dde_new <- add_dea(dde, new_del)
dde_new

# removing DEAs
dde_removed <- remove_dea(dde, "ifng_vs_naive")
dde_removed
}
