% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/DeeDeeExperiment.R
\docType{class}
\name{DeeDeeExperiment-class}
\alias{DeeDeeExperiment-class}
\alias{DeeDeeExperiment}
\title{DeeDeeExperiment}
\usage{
DeeDeeExperiment(se = NULL, de_results = NULL)
}
\arguments{
\item{se}{SumExp}

\item{de_results}{ssss}
}
\value{
tODO
}
\description{
dde <- DeeDeeExperiment(se_macrophage, de_results = del)
}
\section{Slots}{

\describe{
\item{\code{dea}}{todo}
}}

\examples{
data("de_named_list", package = "DeeDee")

dde_onlyde <- DeeDeeExperiment(
  de_results = del
)

# or, with a SE object as support - even without assay data available
library("SummarizedExperiment")

rd_macrophage <- DataFrame(
  gene_id = rownames(del$ifng_vs_naive))
rownames(rd_macrophage) <- rownames(del$ifng_vs_naive)
se_macrophage_noassays <- SummarizedExperiment(
  assays = SimpleList(),
  rowData = rd_macrophage
)

dde <- DeeDeeExperiment(
  se_macrophage_noassays,
  de_results = del
)
}
