% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeeDeeObject.R
\docType{class}
\name{DeeDeeObject-class}
\alias{DeeDeeObject-class}
\alias{DeeDeeObject}
\title{DeeDeeObject (S4 class)}
\description{
DeeDeeObject is the format that the DeeDee App works on. It
holds a slot for your list of DeeDee tables. More functionality
to come.
}
\section{Slots}{

\describe{
\item{\code{DeeDeeList}}{}
}}

\examples{

data(DE_results_IFNg_naive, package = "DeeDee")
IFNg_naive <- deedee_prepare(IFNg_naive, "DESeq2")

data(DE_results_IFNg_both, package = "DeeDee")
IFNg_both <- deedee_prepare(IFNg_both, "DESeq2")

data(DE_results_Salm_naive, package = "DeeDee")
Salm_naive <- deedee_prepare(Salm_naive, "DESeq2")

data(DE_results_Salm_both, package = "DeeDee")
Salm_both <- deedee_prepare(Salm_both, "DESeq2")

dd_list <- list(
  IFNg_naive = IFNg_naive, IFNg_both = IFNg_both,
  Salm_naive = Salm_naive, Salm_both = Salm_both
)

# obj <- DeeDeeObject(DeeDeeList = dd_list)

}
