% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeeDeeExperiment.R
\name{get_deas_list}
\alias{get_deas_list}
\title{this reminds the format of the DeeDee list object}
\usage{
get_deas_list(dde)
}
\arguments{
\item{dde}{todo}
}
\value{
todo
}
\description{
this reminds the format of the DeeDee list object
}
\examples{
# todo
}
