% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeeDeeExperiment.R
\name{remove_dea}
\alias{remove_dea}
\title{Title}
\usage{
remove_dea(dde, dea_name)
}
\arguments{
\item{dde}{todo}

\item{dea_name}{todo}
}
\value{
todo
}
\description{
Title
}
\examples{
data("de_named_list", package = "DeeDee")
library("SummarizedExperiment")

rd_macrophage <- DataFrame(
  gene_id = rownames(del$ifng_vs_naive))
rownames(rd_macrophage) <- rownames(del$ifng_vs_naive)
se_macrophage_noassays <- SummarizedExperiment(
  assays = SimpleList(),
  rowData = rd_macrophage
)
dde <- DeeDeeExperiment(
  se_macrophage_noassays,
  de_results = del
)
dde

dde_new <- remove_dea(dde, "ifng_vs_naive")
}
