% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeeDeeExperiment.R
\name{add_dea}
\alias{add_dea}
\title{dde : the deedee exp object
dea: the named list, as for the main constructor
Title}
\usage{
add_dea(dde, dea)
}
\arguments{
\item{dde}{todo}

\item{dea}{todo}
}
\value{
todo
}
\description{
dde : the deedee exp object
dea: the named list, as for the main constructor
Title
}
\examples{
data("de_named_list", package = "DeeDee")
library("SummarizedExperiment")

rd_macrophage <- DataFrame(
  gene_id = rownames(del$ifng_vs_naive))
rownames(rd_macrophage) <- rownames(del$ifng_vs_naive)
se_macrophage_noassays <- SummarizedExperiment(
  assays = SimpleList(),
  rowData = rd_macrophage
)
dde <- DeeDeeExperiment(
  se_macrophage_noassays,
  de_results = del
)
dde

new_del <- list(
  ifng2 = del$ifng_vs_naive,
  ifngsalmo2 = del$ifngsalmo_vs_naive
)
# add a new (set of) DE result(s)
dde_new <- add_dea(dde, new_del)
}
