% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeeDee_funcs.R
\name{ddedde_upset}
\alias{ddedde_upset}
\title{deedee_upset(dd_list_original, pthresh = 0.05, mode = "both_colored", min_setsize = 10)
Title}
\usage{
ddedde_upset(dde, mode = "both_colored", min_setsize = 10, pthresh = 0.05)
}
\arguments{
\item{dde}{todo}

\item{mode}{todo}

\item{min_setsize}{todo}

\item{pthresh}{todo}
}
\value{
todo
}
\description{
deedee_upset(dd_list_original, pthresh = 0.05, mode = "both_colored", min_setsize = 10)
Title
}
\examples{
data("de_named_list", package = "DeeDee")
library("SummarizedExperiment")

rd_macrophage <- DataFrame(
  gene_id = rownames(del$ifng_vs_naive))
rownames(rd_macrophage) <- rownames(del$ifng_vs_naive)
se_macrophage_noassays <- SummarizedExperiment(
  assays = SimpleList(),
  rowData = rd_macrophage
)
dde <- DeeDeeExperiment(
  se_macrophage_noassays,
  del
)
dde
}
