% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeeDee-pkg.R
\docType{package}
\name{DeeDee-pkg}
\alias{DeeDee-pkg}
\title{DeeDee}
\description{
\code{DeeDee} is a package that allows for the visual comparison of two or more
DEA results.
}
\details{
\code{DeeDee} contains functions for the preprocessing of raw DEA result data and
multiple functions that create graphical output and give an impression of
similarities and differences in the expression profiles. The functions are
contained individually as well as included in an interactive Shiny Web App.
}
