% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_app.R
\name{deedee_app}
\alias{deedee_app}
\title{DeeDee App}
\usage{
deedee_app(deedee_obj = NULL)
}
\arguments{
\item{deedee_obj}{An object of the class DeeDeeObject to be analyzed.}
}
\value{
A shiny app
}
\description{
\code{deedee_app} opens the DeeDee Shiny web application, combining
the functionalities of all other DeeDee functions with a user-friendly
graphical user interface.
}
\examples{

data(dde_macrophage, package = "DeeDee")
deedee_app(dde_macrophage)

}
