% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_app.R
\name{deedee_app}
\alias{deedee_app}
\title{DeeDee App}
\usage{
deedee_app()
}
\value{
A shiny app
}
\description{
\code{deedee_app} opens the DeeDee Shiny web application, combining
the functionalities of all other DeeDee functions with a user-friendly
graphical user interface.
}
\examples{
if (interactive()) {
  deedee_app()
}
}
