% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_deedee_app.R
\name{deedee_app}
\alias{deedee_app}
\title{DeeDee App}
\usage{
deedee_app(deedee_obj = NULL)
}
\arguments{
\item{deedee_obj}{An object of the class DeeDeeObject to be analyzed.}
}
\value{
A shiny app
}
\description{
\code{deedee_app} opens the DeeDee Shiny web application, combining
the functionalities of all other DeeDee functions with a user-friendly
graphical user interface.
}
\examples{

data(DE_results_IFNg_naive, package = "DeeDee")
IFNg_naive <- deedee_prepare(IFNg_naive, "DESeq2")

data(DE_results_IFNg_both, package = "DeeDee")
IFNg_both <- deedee_prepare(IFNg_both, "DESeq2")

data(DE_results_Salm_naive, package = "DeeDee")
Salm_naive <- deedee_prepare(Salm_naive, "DESeq2")

data(DE_results_Salm_both, package = "DeeDee")
Salm_both <- deedee_prepare(Salm_both, "DESeq2")

dd_list <- list(
  IFNg_naive = IFNg_naive, IFNg_both = IFNg_both,
  Salm_naive = Salm_naive, Salm_both = Salm_both
)

if (interactive()) {
  deedee_app(dd_list)
}
}
