% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_app_old.R
\name{deedee_app_deprecated}
\alias{deedee_app_deprecated}
\title{DeeDee App}
\usage{
deedee_app_deprecated(deedee_obj = NULL)
}
\arguments{
\item{deedee_obj}{An object of the class DeeDeeObject to be analyzed.}
}
\value{
A shiny app
}
\description{
\code{deedee_app} opens the DeeDee Shiny web application, combining
the functionalities of all other DeeDee functions with a user-friendly
graphical user interface.
}
