% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeeDee_cat.R
\name{deedee_cat}
\alias{deedee_cat}
\title{deedee_cat}
\usage{
deedee_cat(dde, ref = 1, maxrank = 1000, mode = "up", pthresh = 0.05)
}
\arguments{
\item{dde}{A \link{DeeDeeExperiment} object.}

\item{ref}{A numeric value, corresponding to the order of the element in the
\code{dde} object to be used as a reference.}

\item{maxrank}{A numeric value. Indicates the maximum ranked feature to
include when computing the Concordance At the Top. Defaults to 1000.}

\item{mode}{A character value, could be one of "up", "down", or "both". Defines
which features to include in the computations. Defaults to "up".}

\item{pthresh}{Numeric value, corresponding to the p-value to use as a
threshold to subset the features to include. Defaults sensibly to 0.05.}
}
\value{
A \code{ggplot} plot object.
}
\description{
CAT plot - Concordance At the Top
}
\examples{
data("de_named_list", package = "DeeDee")
library("SummarizedExperiment")

rd_macrophage <- DataFrame(
  gene_id = rownames(de_named_list$ifng_vs_naive))
rownames(rd_macrophage) <- rownames(de_named_list$ifng_vs_naive)
se_macrophage_noassays <- SummarizedExperiment(
  assays = SimpleList(),
  rowData = rd_macrophage
)
dde <- DeeDeeExperiment(
  se_macrophage_noassays,
  de_results = de_named_list
)
dde

deedee_cat(dde, ref = 1, maxrank = 1000, mode = "up", pthresh = 0.05)

}
