% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_cat.R
\name{deedee_cat}
\alias{deedee_cat}
\title{Title}
\usage{
deedee_cat(data, pthresh = 0.05, maxrank = 1000)
}
\arguments{
\item{data}{(named) list of results from deedee_prepare}

\item{pthresh}{threshold for p-values to be in-/excluded (default = 0.05}

\item{maxrank}{highest rank that should be displayed (default = 1000)}
}
\value{
ggplot object (plottable with show()/print())
}
\description{
"deedee_cat" creates a Concordance at the Top plot depicting
the concordance of genes in the first n elements of the logFC-ranked lists
for the given input datasets.
}
\examples{

}
