% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeeDee_cat.R
\name{deedee_cat}
\alias{deedee_cat}
\title{Title}
\usage{
deedee_cat(dde, ref = 1, maxrank = 1000, mode = "up", pthresh = 0.05)
}
\arguments{
\item{dde}{todo}

\item{ref}{todo}

\item{maxrank}{todo}

\item{mode}{todo}

\item{pthresh}{todo}
}
\value{
todo
}
\description{
Title
}
\examples{
data("de_named_list", package = "DeeDee")
library("SummarizedExperiment")

rd_macrophage <- DataFrame(
  gene_id = rownames(del$ifng_vs_naive))
rownames(rd_macrophage) <- rownames(del$ifng_vs_naive)
se_macrophage_noassays <- SummarizedExperiment(
  assays = SimpleList(),
  rowData = rd_macrophage
)
dde <- DeeDeeExperiment(
  se_macrophage_noassays,
  de_results = del
)
dde

deedee_cat(dde, ref = 1, maxrank = 1000, mode = "up", pthresh = 0.05)
}
