% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_cat.R
\name{deedee_cat}
\alias{deedee_cat}
\title{DeeDee Concordance At the Top Plot}
\usage{
deedee_cat(data, ref = 1, maxrank = 1000, mode = "up", pthresh = 0.05)
}
\arguments{
\item{data}{(named) list of results from deedee_prepare}

\item{ref}{index of the contrast in data to be used as reference contrast
(default = 1)}

\item{maxrank}{highest rank that should be displayed (default = 1000)}

\item{mode}{sort by highest logFC (\code{up}, default), lowest logFC (\code{down}) or
greatest deviation from zero (\code{both})}

\item{pthresh}{threshold for p-values to be in-/excluded (default = 0.05}
}
\value{
ggplot object (plottable with show()/print())
}
\description{
\code{deedee_cat} creates a Concordance at the Top plot depicting
the concordance of genes in the first n elements of the logFC-ranked lists
for the given input datasets.
}
\examples{

data(DE_results_IFNg_naive, package = "DeeDee")
IFNg_naive <- deedee_prepare(IFNg_naive, "DESeq2")

data(DE_results_IFNg_both, package = "DeeDee")
IFNg_both <- deedee_prepare(IFNg_both, "DESeq2")

data(DE_results_Salm_naive, package = "DeeDee")
Salm_naive <- deedee_prepare(Salm_naive, "DESeq2")

data(DE_results_Salm_both, package = "DeeDee")
Salm_both <- deedee_prepare(Salm_both, "DESeq2")

dd_list <- list(
  IFNg_naive = IFNg_naive, IFNg_both = IFNg_both,
  Salm_naive = Salm_naive, Salm_both = Salm_both
)

deedee_cat(dd_list, ref = 1, maxrank = 1000, mode = "up", pthresh = 0.05)
}
