% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_from_dde.R
\name{deedee_from_dde}
\alias{deedee_from_dde}
\title{DeeDee Prepare}
\usage{
deedee_from_dde(dde)
}
\arguments{
\item{dde}{Instance of the DeeDeeExperiment class}
}
\value{
list of DeeDee tables, to be used as part of the input for the other
DeeDee functions
}
\description{
\code{deedee_from_dde} prepares an instance of the DeeDeeExperiment
class for the usage in other DeeDee functions.
}
\examples{

deedee_table <- deedee_from_dde(dde)

}
