% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeeDee_heatmap.R
\name{deedee_heatmap}
\alias{deedee_heatmap}
\title{deedee_heatmap}
\usage{
deedee_heatmap(
  dde,
  show_first = 25,
  show_gene_names = FALSE,
  dist = "euclidean",
  clust = "average",
  show_na = FALSE,
  pthresh = 0.05
)
}
\arguments{
\item{dde}{A \link{DeeDeeExperiment} object.}

\item{show_first}{Numeric value, specifying the number of features to include.
The ranking is done according to ... TODO}

\item{show_gene_names}{Logical value, whether to display the gene names on the
heatmap's side.}

\item{dist}{Character value, specifying the distance type to use in the call
to \code{ComplexHeatmap}. Accordingly, it can be a pre-defined character
("euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski",
"pearson", "spearman", "kendall").}

\item{clust}{Character value. Defines the method to perform hierarchical
clustering, passed to hclust, as used in \code{ComplexHeatmap}.}

\item{show_na}{Logical value, whether to include features that have NA value
for the log fold change.}

\item{pthresh}{Numeric value, corresponding to the p-value to use as a
threshold to subset the features to include.}
}
\value{
A plot object generated via the \code{ComplexHeatmap} package.
}
\description{
Construct a logFC heatmap on a subset of features
}
\examples{
data("de_named_list", package = "DeeDee")
library("SummarizedExperiment")

rd_macrophage <- DataFrame(
  gene_id = rownames(del$ifng_vs_naive))
rownames(rd_macrophage) <- rownames(del$ifng_vs_naive)
se_macrophage_noassays <- SummarizedExperiment(
  assays = SimpleList(),
  rowData = rd_macrophage
)
dde <- DeeDeeExperiment(
  se_macrophage_noassays,
  de_results = del
)
dde

deedee_heatmap(dde,
               pthresh = 0.05, show_first = 25,
               show_gene_names = FALSE, dist = "euclidean",
               clust = "average", show_na = FALSE
)

}
