% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeeDee_heatmap.R
\name{deedee_heatmap}
\alias{deedee_heatmap}
\title{Title}
\usage{
deedee_heatmap(
  dde,
  show_first = 25,
  show_gene_names = FALSE,
  dist = "euclidean",
  clust = "average",
  show_na = FALSE,
  pthresh = 0.05
)
}
\arguments{
\item{dde}{todo}

\item{show_first}{todo}

\item{show_gene_names}{todo}

\item{dist}{todo}

\item{clust}{todo}

\item{show_na}{todo}

\item{pthresh}{todo}
}
\value{
todo
}
\description{
Title
}
\examples{
data("de_named_list", package = "DeeDee")
library("SummarizedExperiment")

rd_macrophage <- DataFrame(
  gene_id = rownames(del$ifng_vs_naive))
rownames(rd_macrophage) <- rownames(del$ifng_vs_naive)
se_macrophage_noassays <- SummarizedExperiment(
  assays = SimpleList(),
  rowData = rd_macrophage
)
dde <- DeeDeeExperiment(
  se_macrophage_noassays,
  de_results = del
)
dde

deedee_heatmap(dde,
               pthresh = 0.05, show_first = 25,
               show_gene_names = FALSE, dist = "euclidean",
               clust = "average", show_na = FALSE
)

}
