% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_deedee_heatmap.R
\name{deedee_heatmap}
\alias{deedee_heatmap}
\title{DeeDee Heatmap}
\usage{
deedee_heatmap(
  data,
  show_first = 25,
  show_gene_names = FALSE,
  dist = "euclidean",
  clust = "average",
  show_na = FALSE,
  pthresh = 0.05
)
}
\arguments{
\item{data}{named list of results from deedee_prepare()}

\item{show_first}{indicating the number of genes depicted (default = 25)}

\item{show_gene_names}{boolean, show row names next to heatmap
(default = FALSE)}

\item{dist}{select the distance measure (\code{euclidean}, \code{manhattan}, \code{pearson},
\code{spearman})}

\item{clust}{select the clustering method (\code{single}, \code{complete}, \code{average},
\code{centroid})}

\item{show_na}{boolean, include genes with NA values in heatmap
(default = FALSE)}

\item{pthresh}{threshold for p-values to be in-/excluded (default = 0.05)}
}
\value{
Heatmap object (plottable with show()/print()). The resulting heatmap
(\code{res}) can be opened in a Shiny App with interactive functionality by
running the command \code{ht_shiny(res)} (see package \code{InteractiveComplexHeatmap}).
}
\description{
\code{deedee_heatmap} creates a heatmap depicting the logFC as a
measure of the differential expression of the first genes in the given
datasets.
}
\examples{

data(DE_results_IFNg_naive, package = "DeeDee")
IFNg_naive <- deedee_prepare(IFNg_naive, "DESeq2")

data(DE_results_IFNg_both, package = "DeeDee")
IFNg_both <- deedee_prepare(IFNg_both, "DESeq2")

data(DE_results_Salm_naive, package = "DeeDee")
Salm_naive <- deedee_prepare(Salm_naive, "DESeq2")

data(DE_results_Salm_both, package = "DeeDee")
Salm_both <- deedee_prepare(Salm_both, "DESeq2")

dd_list <- list(
  IFNg_naive = IFNg_naive, IFNg_both = IFNg_both,
  Salm_naive = Salm_naive, Salm_both = Salm_both
)

deedee_heatmap(dd_list,
  pthresh = 0.05, show_first = 25,
  show_gene_names = FALSE, dist = "euclidean",
  clust = "average", show_na = FALSE
)
}
