% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_heatmap.R
\name{deedee_heatmap}
\alias{deedee_heatmap}
\title{DeeDee Heatmap}
\usage{
deedee_heatmap(
  data,
  pthresh = 0.05,
  show_first = 25,
  show_gene_names = FALSE,
  dist = "euclidean",
  clust = "average",
  show_na = FALSE
)
}
\arguments{
\item{data}{(named) list of 2-5 results from deedee_format}

\item{pthresh}{threshold for p-values to be in-/excluded (default = 0.05)}

\item{show_first}{indicating the number of genes depicted (default = 500)}

\item{show_gene_names}{boolean, show rownames next to heatmap
(default = FALSE)}
}
\value{
ggplot object (plottable with show()/print())
}
\description{
"deedee_heatmap" creates a heatmap depicting the logFC as a
measure of the differential expression of the first genes in the given
datasets.
}
\examples{

}
