% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_prepare.R
\name{deedee_prepare}
\alias{deedee_prepare}
\title{DeeDee Prepare}
\usage{
deedee_prepare(data, input_type)
}
\arguments{
\item{data}{result from DESeq2, limma or edgeR}

\item{input_type}{the program the data comes from ("DESeq2", "limma" or
"edgeR")}
}
\value{
table to be used as input for other DeeDee functions
}
\description{
create DeeDee object from DEA result
}
\examples{
library(DeeDee)

data(DE_results_IFNg_naive, package = "DeeDee")

dd_table <- deedee_prepare(IFNg_naive, "DESeq2")
}
