% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_prepare.R
\name{deedee_prepare}
\alias{deedee_prepare}
\title{DeeDee Prepare}
\usage{
deedee_prepare(data, input_type)
}
\arguments{
\item{data}{result from DESeq2, limma or edgeR}

\item{input_type}{the program the data comes from (\code{DESeq2}, \code{limma} or
\code{edgeR})}
}
\value{
Dee Dee table, to be used as part of the input for the other DeeDee
functions
}
\description{
\code{deedee_prepare} creates a DeeDee table from a DEA result
}
\examples{

data(DE_results_IFNg_naive, package = "DeeDee")

deedee_table <- deedee_prepare(IFNg_naive, "DESeq2")
}
