% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_prepare.R
\name{deedee_prepare}
\alias{deedee_prepare}
\title{DeeDee Prepare}
\usage{
deedee_prepare(data, input_type)
}
\arguments{
\item{data}{result from DESeq2, limma or edgeR}

\item{input_type}{the program the data comes from ("DESeq2", "limma" or
"edgeR")}
}
\value{
table to be used as input for other DeeDee functions
}
\description{
create DeeDee object from DEA result
}
\examples{
  library(DeeDee)

  library("macrophage")
  library("DESeq2")
  source("~/Development/DeeDee_wip/deedee_prepare.R")

  data(gse, "macrophage")

  dds_macrophage <- DESeqDataSet(gse, design = ~line + condition)
  rownames(dds_macrophage) <- substr(rownames(dds_macrophage), 1, 15)

  keep <- rowSums(counts(dds_macrophage) >= 10) >= 6
  dds_macrophage <- dds_macrophage[keep, ]
  dds_macrophage <- DESeq(dds_macrophage)

  IFNg_naive <- results(dds_macrophage,
  contrast = c("condition", "IFNg", "naive"),
  lfcThreshold = 1, alpha = 0.05)

  dd_table <- deedee_prepare(IFNg_naive, "DESeq2")

}
