% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_deedee_qq.R
\name{deedee_qq}
\alias{deedee_qq}
\title{DeeDee QQ Plot}
\usage{
deedee_qq(
  data,
  select1 = 1,
  select2 = 2,
  color_by = "pval1",
  as_line = FALSE,
  pthresh = 0.05
)
}
\arguments{
\item{data}{named list of results from deedee_prepare()}

\item{select1}{index of first data-list element to be used (default = 1)}

\item{select2}{index of second data-list element to be used (default = 2)}

\item{color_by}{indicates which set of values the output should be colored by
(possible values = \code{pval1} (default), \code{pval2})}

\item{as_line}{logical value specifying if the resulting plot should be a
line (TRUE) or points (FALSE, default)}

\item{pthresh}{threshold for p-values to be in-/excluded (default = 0.05)}
}
\value{
ggplot object (plottable with show()/print())
}
\description{
\code{deedee_qq} creates a Q-Q-plot comparing the statistical
distribution of the logFC of the genes in two input datasets. To compare more
than two contrasts against a reference, have a look at \code{deedee_qqmult}.
}
\examples{

data(DE_results_IFNg_naive, package = "DeeDee")
IFNg_naive <- deedee_prepare(IFNg_naive, "DESeq2")

data(DE_results_IFNg_both, package = "DeeDee")
IFNg_both <- deedee_prepare(IFNg_both, "DESeq2")

data(DE_results_Salm_naive, package = "DeeDee")
Salm_naive <- deedee_prepare(Salm_naive, "DESeq2")

data(DE_results_Salm_both, package = "DeeDee")
Salm_both <- deedee_prepare(Salm_both, "DESeq2")

dd_list <- list(
  IFNg_naive = IFNg_naive, IFNg_both = IFNg_both,
  Salm_naive = Salm_naive, Salm_both = Salm_both
)

deedee_qq(dd_list, pthresh = 0.05, select = 1, select2 = 2, color_by = "pval1")
}
