% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeeDee_qq.R
\name{deedee_qq}
\alias{deedee_qq}
\title{deedee_qq}
\usage{
deedee_qq(
  dde,
  select1 = 1,
  select2 = 2,
  color_by = "pval1",
  as_line = FALSE,
  pthresh = 0.05
)
}
\arguments{
\item{dde}{A \link{DeeDeeExperiment} object.}

\item{select1}{todo}

\item{select2}{todo}

\item{color_by}{todo}

\item{as_line}{todo}

\item{pthresh}{Numeric value, corresponding to the p-value to use as a
threshold to subset the features to include.}
}
\value{
A \code{ggplot} plot object.
}
\description{
Q-Q plot for comparing DE analyses
}
\examples{
data("de_named_list", package = "DeeDee")
library("SummarizedExperiment")

rd_macrophage <- DataFrame(
  gene_id = rownames(del$ifng_vs_naive))
rownames(rd_macrophage) <- rownames(del$ifng_vs_naive)
se_macrophage_noassays <- SummarizedExperiment(
  assays = SimpleList(),
  rowData = rd_macrophage
)
dde <- DeeDeeExperiment(
  se_macrophage_noassays,
  de_results = del
)
dde

deedee_qq(dde)

}
