% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_qq.R
\name{deedee_qq}
\alias{deedee_qq}
\title{DeeDee QQ Plot}
\usage{
deedee_qq(
  data,
  pthresh = 0.05,
  select1 = 1,
  select2 = 2,
  color_by = "pval1",
  ggplot = FALSE
)
}
\arguments{
\item{data}{(named) list of results from deedee_prepare}

\item{pthresh}{threshold for p-values to be in-/excluded (default = 0.05)}

\item{select1}{index of first data-list element to be used (default = 1)}

\item{select2}{index of second data-list element to be used (default = 2)}

\item{color_by}{indicates which set of values the output should be colored by
(possible values = "logFC1", "logFC2", "pval1" (default),
"pval2")}

\item{ggplot}{output as ggplot (default = FALSE)}
}
\value{
ggplot object (plottable with show()/print())
}
\description{
"deedee_qq" creates a QQ-plot comparing the statistical
distribution of the logFC of the genes in two input datasets.
}
\examples{

}
