% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_qq.R
\name{deedee_qq}
\alias{deedee_qq}
\title{DeeDee QQ Plot}
\usage{
deedee_qq(data, pthresh = 0.05, select1 = 1, select2 = 2, color_by = "pval1")
}
\arguments{
\item{data}{(named) list of results from deedee_prepare}

\item{pthresh}{threshold for p-values to be in-/excluded (default = 0.05)}

\item{select1}{index of first data-list element to be used (default = 1)}

\item{select2}{index of second data-list element to be used (default = 2)}

\item{color_by}{indicates which set of values the output should be colored by
(possible values = "pval1" (default),
"pval2")}
}
\value{
ggplot object (plottable with show()/print())
}
\description{
"deedee_qq" creates a QQ-plot comparing the statistical
distribution of the logFC of the genes in two input datasets.
}
\examples{
  library("macrophage")
  library("DESeq2")
  source("~/Development/DeeDee_wip/deedee_prepare.R")

  data(gse, "macrophage")

  dds_macrophage <- DESeqDataSet(gse, design = ~line + condition)
  rownames(dds_macrophage) <- substr(rownames(dds_macrophage), 1, 15)

  keep <- rowSums(counts(dds_macrophage) >= 10) >= 6
  dds_macrophage <- dds_macrophage[keep, ]
  dds_macrophage <- DESeq(dds_macrophage)

  IFNg_naive <- results(dds_macrophage,
  contrast = c("condition", "IFNg", "naive"),
  lfcThreshold = 1, alpha = 0.05)
  inp1 <- deedee_prepare(IFNg_naive, "DESeq2")

  IFNg_IFNgSL1344 <- results(dds_macrophage,
  contrast = c("condition", "IFNg_SL1344", "IFNg"),
  lfcThreshold = 1, alpha = 0.05)
  inp2 <- deedee_prepare(IFNg_IFNgSL1344, "DESeq2")

 SL1344_naive <- results(dds_macrophage,
 contrast = c("condition", "SL1344", "naive"),
 lfcThreshold = 1, alpha = 0.05)
 inp3 <- deedee_prepare(SL1344_naive, "DESeq2")

 SL1344_IFNgSL1344 <- results(dds_macrophage,
 contrast = c("condition", "IFNg_SL1344", "SL1344"),
 lfcThreshold = 1, alpha = 0.05)
 inp4 <- deedee_prepare(SL1344_IFNgSL1344, "DESeq2")

 dd_list <- list(naive_IFNg = inp1, IFNg_both = inp2, naive_Salm = inp3,
 Salm_both = inp4)

 deedee_qq(dd_list, pthresh = 0.05, select = 1, select2 = 2, color_by = "pval1")

}
