% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_qqmult.R
\name{deedee_qqmult}
\alias{deedee_qqmult}
\title{DeeDee QQ Plot with Multiple Contrasts}
\usage{
deedee_qqmult(data, ref = 1, pthresh = 0.05)
}
\arguments{
\item{data}{instance of the DeeDeeExperiment class}

\item{ref}{index of the contrast in data to be used as reference contrast
(default = 1)}

\item{pthresh}{threshold for p-values to be in-/excluded (default = 0.05)}
}
\value{
ggplot object (plottable with show()/print())
}
\description{
\code{deedee_qqmult} creates a plot containing Q-Q-lines comparing
the statistical distribution of the logFC of the genes in each input datasets
to a contrast chosen from the input data.
}
\examples{

data(dde_macrophage, package = "DeeDee")
deedee_qqmult(dde_macrophage)

}
