% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeeDee_qqmult.R
\name{deedee_qqmult}
\alias{deedee_qqmult}
\title{deedee_qqmult}
\usage{
deedee_qqmult(dde, ref = 1, pthresh = 0.05)
}
\arguments{
\item{dde}{A \link{DeeDeeExperiment} object.}

\item{ref}{A numeric value, corresponding to the order of the element in the
\code{dde} object to be used as a reference.}

\item{pthresh}{Numeric value, corresponding to the p-value to use as a
threshold to subset the features to include.}
}
\value{
A \code{ggplot} plot object.
}
\description{
Q-Q plot for multiple comparisons of DE analyses
}
\examples{
data("de_named_list", package = "DeeDee")
library("SummarizedExperiment")

rd_macrophage <- DataFrame(
  gene_id = rownames(de_named_list$ifng_vs_naive))
rownames(rd_macrophage) <- rownames(de_named_list$ifng_vs_naive)
se_macrophage_noassays <- SummarizedExperiment(
  assays = SimpleList(),
  rowData = rd_macrophage
)
dde <- DeeDeeExperiment(
  se_macrophage_noassays,
  de_results = de_named_list
)
dde

deedee_qqmult(dde, pthresh = 0.05, ref = 1)

}
