% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_rrho.R
\name{deedee_rrho}
\alias{deedee_rrho}
\title{DeeDee RRHO Plot}
\usage{
deedee_rrho(
  data,
  select1 = 1,
  select2 = 2,
  alternative = "two.sided",
  corr = FALSE
)
}
\arguments{
\item{data}{instance of the DeeDeeExperiment class}

\item{select1}{index of the 1st contrast to be compared (default = 1)}

\item{select2}{index of the 2nd contrast to be compared (default = 2)}

\item{alternative}{\code{"two.sided"} (default) for concordant and discordant
regions, or \code{"enrichment"} for a one-sided test}

\item{corr}{boolean, apply Benjamini-Yekutieli FDR correction to the map
(default = FALSE)}
}
\value{
plottable RRHO object
}
\description{
DeeDee RRHO Plot
}
\examples{

data(dde_macrophage, package = "DeeDee")
deedee_rrho(dde_macrophage, select1 = 1, select2 = 2)

}
