% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_scatter.R
\name{deedee_scatter}
\alias{deedee_scatter}
\title{DeeDee logFC Scatter Plot}
\usage{
deedee_scatter(
  data,
  select1 = 1,
  select2 = 2,
  color_by = "pval1",
  pthresh = 0.05
)
}
\arguments{
\item{data}{instance of the DeeDeeExperiment class}

\item{select1}{index of first data-list element to be used (default = 1)}

\item{select2}{index of second data-list element to be used (default = 2)}

\item{color_by}{indicates which set of values the output should be colored by
(possible values = \code{pval1} (default), \code{pval2})}

\item{pthresh}{threshold for p-values to be in-/excluded (default = 0.05)}
}
\value{
ggplot object (plottable with show()/print())
}
\description{
\code{deedee_scatter} creates a scatterplot of the genes in two input
datasets based on their logFC values.
}
\examples{

data(dde_macrophage, package = "DeeDee")
deedee_scatter(dde_macrophage, select1 = 1, select2 = 2)

}
