% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_scatter.R
\name{deedee_scatter}
\alias{deedee_scatter}
\title{DeeDee logFC Scatterplot}
\usage{
deedee_scatter(
  data,
  pthresh = 0.05,
  select1 = 1,
  select2 = 2,
  color_by = "pval1"
)
}
\arguments{
\item{data}{list of results from deedee_prepare}

\item{pthresh}{threshold for p-values to be in-/excluded (default = 0.05)}

\item{select1}{index of first data-list element to be used (default = 1)}

\item{select2}{index of second data-list element to be used (default = 2)}

\item{color_by}{indicates which set of values the output should be colored by
(possible values = "pval1" (default), "pval2", "pval_mean",
"idr") (blue -> lower values, red -> higher values)}
}
\value{
ggplot object (plottable with show()/print())
}
\description{
"deedee_scatter" creates a scatterplot of the genes in two input
datasets based on their logFC values.
}
\examples{

}
