% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_scatter.R
\name{deedee_scatter}
\alias{deedee_scatter}
\title{DeeDee logFC Scatter Plot}
\usage{
deedee_scatter(
  data,
  pthresh = 0.05,
  select1 = 1,
  select2 = 2,
  color_by = "pval1"
)
}
\arguments{
\item{data}{list of results from deedee_prepare}

\item{pthresh}{threshold for p-values to be in-/excluded (default = 0.05)}

\item{select1}{index of first data-list element to be used (default = 1)}

\item{select2}{index of second data-list element to be used (default = 2)}

\item{color_by}{indicates which set of values the output should be colored by
(possible values = "pval1" (default), "pval2", "idr") (blue
-> lower values, red -> higher values)}
}
\value{
ggplot object (plottable with show()/print())
}
\description{
"deedee_scatter" creates a scatterplot of the genes in two input
datasets based on their logFC values.
}
\examples{
library(DeeDee)

data(DE_results_IFNg_naive, package = "DeeDee")
IFNg_naive <- deedee_prepare(IFNg_naive, "DESeq2")

data(DE_results_IFNg_both, package = "DeeDee")
IFNg_both <- deedee_prepare(IFNg_both, "DESeq2")

data(DE_results_Salm_naive, package = "DeeDee")
Salm_naive <- deedee_prepare(Salm_naive, "DESeq2")

data(DE_results_Salm_both, package = "DeeDee")
Salm_both <- deedee_prepare(Salm_both, "DESeq2")

dd_list <- list(
  IFNg_naive = IFNg_naive, IFNg_both = IFNg_both,
  Salm_naive = Salm_naive, Salm_both = Salm_both
)

deedee_scatter(dd_list,
  pthresh = 0.05, select1 = 1, select2 = 2,
  color_by = "pval1"
)
}
