% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeeDee_upset.R
\name{deedee_upset}
\alias{deedee_upset}
\title{deedee_upset}
\usage{
deedee_upset(dde, mode = "both_colored", min_setsize = 10, pthresh = 0.05)
}
\arguments{
\item{dde}{A \link{DeeDeeExperiment} object.}

\item{mode}{Character value, specifies which subset of features to include
in the overlap computations. Can be either of the following: "both", "up",
"down", or "both_colored".}

\item{min_setsize}{Numeric value, specifying the minimal number of
observations in an intersection for it to be included in the upset plot.}

\item{pthresh}{Numeric value, corresponding to the p-value to use as a
threshold to subset the features to include.}
}
\value{
A plot object, drawn with the \code{ComplexUpset} package.
}
\description{
Upset plot on the sets of DE features for the different analyses included
}
\examples{
data("de_named_list", package = "DeeDee")
library("SummarizedExperiment")

rd_macrophage <- DataFrame(
  gene_id = rownames(de_named_list$ifng_vs_naive))
rownames(rd_macrophage) <- rownames(de_named_list$ifng_vs_naive)
se_macrophage_noassays <- SummarizedExperiment(
  assays = SimpleList(),
  rowData = rd_macrophage
)
dde <- DeeDeeExperiment(
  se_macrophage_noassays,
  de_results = de_named_list
)
dde

deedee_upset(dde, pthresh = 0.05, mode = "both_colored", min_setsize = 10)

}
