% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_upset.R
\name{deedee_upset}
\alias{deedee_upset}
\title{DeeDee UpSet Plot}
\usage{
deedee_upset(data, pthresh = 0.05, mode = "both_colored", min_setsize = 10)
}
\arguments{
\item{data}{list of results from deedee_prepare}

\item{pthresh}{threshold for p-values to be in-/excluded (default = 0.05)}

\item{mode}{show all overlapping DE genes (\code{both}, default),
all overlapping genes colored by DE direction (\code{both_colored},
red -> all up, blue -> all down),
only conjointly up-regulated (\code{up})
or only conjointly down-regulated (\code{down}) genes}

\item{min_setsize}{the minimum size of intersections to be displayed in the
UpSet plot (default = 10)}
}
\value{
upset element (plottable with show()/print())
}
\description{
\code{deedee_upset} creates an UpSet plot depicting the overlaps of
differentially expressed genes in the input datasets.
}
\examples{

data(DE_results_IFNg_naive, package = "DeeDee")
IFNg_naive <- deedee_prepare(IFNg_naive, "DESeq2")

data(DE_results_IFNg_both, package = "DeeDee")
IFNg_both <- deedee_prepare(IFNg_both, "DESeq2")

data(DE_results_Salm_naive, package = "DeeDee")
Salm_naive <- deedee_prepare(Salm_naive, "DESeq2")

data(DE_results_Salm_both, package = "DeeDee")
Salm_both <- deedee_prepare(Salm_both, "DESeq2")

dd_list <- list(
  IFNg_naive = IFNg_naive, IFNg_both = IFNg_both,
  Salm_naive = Salm_naive, Salm_both = Salm_both
)

deedee_upset(dd_list, pthresh = 0.05, mode = "both_colored", min_setsize = 10)
}
