% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_upSet.R
\name{deedee_upSet}
\alias{deedee_upSet}
\title{DeeDee UpSet Plot}
\usage{
deedee_upSet(data, pthresh = 0.05, mode = "both", min_setsize = 10)
}
\arguments{
\item{data}{list of results from deedee_prepare}

\item{pthresh}{threshold for p-values to be in-/excluded (default = 0.05)}

\item{mode}{show all overlapping DE genes ("both", default),
all overlapping genes colored by DE direction ("both_colored",
red -> all up, blue -> all down),
only conjointly up-regulated ("up")
or only conjointly down-regulated ("down") genes}
}
\value{
upset element (plottable with show()/print())
}
\description{
"deedee_upSet" creates an UpSet plot depicting the overlaps of
differentially expressed genes in the input datasets.
}
\examples{

}
