% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_venn.R
\name{deedee_venn}
\alias{deedee_venn}
\title{DeeDee Venn Diagram}
\usage{
deedee_venn(data, pthresh = 0.05, mode = "both")
}
\arguments{
\item{data}{(named) list of 2-4 results from deedee_format}

\item{pthresh}{threshold for p-values to be in-/excluded (default = 0.05)}

\item{mode}{show all overlapping DE genes ("both", default),
only conjointly up-regulated ("up")
or only conjointly down-regulated ("down") genes}
}
\value{
ggplot object (plottable with show()/print())
}
\description{
"deedee_venn" creates a Venn diagram depicting the overlaps of
differentially expressed genes in the input datasets.
}
\examples{
  library("macrophage")
  library("DESeq2")
  source("~/Development/DeeDee_wip/deedee_prepare.R")

  data(gse, "macrophage")

  dds_macrophage <- DESeqDataSet(gse, design = ~line + condition)
  rownames(dds_macrophage) <- substr(rownames(dds_macrophage), 1, 15)

  keep <- rowSums(counts(dds_macrophage) >= 10) >= 6
  dds_macrophage <- dds_macrophage[keep, ]
  dds_macrophage <- DESeq(dds_macrophage)

  IFNg_naive <- results(dds_macrophage,
  contrast = c("condition", "IFNg", "naive"),
  lfcThreshold = 1, alpha = 0.05)
  inp1 <- deedee_prepare(IFNg_naive, "DESeq2")

  IFNg_IFNgSL1344 <- results(dds_macrophage,
  contrast = c("condition", "IFNg_SL1344", "IFNg"),
  lfcThreshold = 1, alpha = 0.05)
  inp2 <- deedee_prepare(IFNg_IFNgSL1344, "DESeq2")

 SL1344_naive <- results(dds_macrophage,
 contrast = c("condition", "SL1344", "naive"),
 lfcThreshold = 1, alpha = 0.05)
 inp3 <- deedee_prepare(SL1344_naive, "DESeq2")

 SL1344_IFNgSL1344 <- results(dds_macrophage,
 contrast = c("condition", "IFNg_SL1344", "SL1344"),
 lfcThreshold = 1, alpha = 0.05)
 inp4 <- deedee_prepare(SL1344_IFNgSL1344, "DESeq2")

 dd_list <- list(naive_IFNg = inp1, IFNg_both = inp2, naive_Salm = inp3,
 Salm_both = inp4)

 deedee_venn(dd_list, pthresh = 0.05, mode = "both")

}
