% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeeDee_venn.R
\name{deedee_venn}
\alias{deedee_venn}
\title{deedee_venn}
\usage{
deedee_venn(dde, mode = "both", pthresh = 0.05)
}
\arguments{
\item{dde}{A \link{DeeDeeExperiment} object.}

\item{mode}{Character value, one of "both", "up", or "down". Specifies which
set of features to focus on for the overlap calculations.}

\item{pthresh}{Numeric value, corresponding to the p-value to use as a
threshold to subset the features to include.}
}
\value{
A \code{ggplot} plot object.
}
\description{
A Venn diagram on the DE features from different analyses
}
\examples{
data("de_named_list", package = "DeeDee")
library("SummarizedExperiment")

rd_macrophage <- DataFrame(
  gene_id = rownames(del$ifng_vs_naive))
rownames(rd_macrophage) <- rownames(del$ifng_vs_naive)
se_macrophage_noassays <- SummarizedExperiment(
  assays = SimpleList(),
  rowData = rd_macrophage
)
dde <- DeeDeeExperiment(
  se_macrophage_noassays,
  de_results = del
)
dde

library("dplyr") ## to have the inner_join, needed by ggvenn
deedee_venn(dde, pthresh = 0.05, mode = "both")

}
