% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_venn.R
\name{deedee_venn}
\alias{deedee_venn}
\title{DeeDee Venn Diagram}
\usage{
deedee_venn(data, pthresh = 0.05, mode = "both")
}
\arguments{
\item{data}{(named) list of 2-4 results from deedee_format}

\item{pthresh}{threshold for p-values to be in-/excluded (default = 0.05)}

\item{mode}{show all overlapping DE genes ("both", default),
only conjointly up-regulated ("up")
or only conjointly down-regulated ("down") genes}
}
\value{
ggplot object (plottable with show()/print())
}
\description{
"deedee_venn" creates a Venn diagram depicting the overlaps of
differentially expressed genes in the input datasets.
}
\examples{

}
