% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeeDeeExperiment-methods.R
\name{DeeDeeExperiment-methods}
\alias{DeeDeeExperiment-methods}
\alias{dea}
\alias{dea<-}
\alias{dea,DeeDeeExperiment-method}
\alias{dea<-,DeeDeeExperiment,list-method}
\title{Methods for DeeDeeExperiment objects}
\usage{
\S4method{dea}{DeeDeeExperiment}(x)

\S4method{dea}{DeeDeeExperiment,list}(x) <- value
}
\arguments{
\item{x}{A \code{\link{DeeDeeExperiment}} object}

\item{value}{Replacement value for replacement methods.}
}
\value{
Return value varies depending on method, as described below.
}
\description{
Todo
}
