% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/DeeDeeExperiment.R
\docType{class}
\name{DeeDeeExperiment-class}
\alias{DeeDeeExperiment-class}
\alias{DeeDeeExperiment}
\title{DeeDeeExperiment}
\usage{
DeeDeeExperiment(..., de_results = NULL)
}
\arguments{
\item{...}{Arguments passed to the \code{\link{SummarizedExperiment}}
constructor to fill the slots of the base class.}

\item{de_results}{ssss}
}
\value{
tODO
}
\description{
dde <- DeeDeeExperiment(se_macrophage, de_results = del)
}
\section{Slots}{

\describe{
\item{\code{dea}}{todo}
}}

\examples{
# todo
}
